Module ovgu_canteen::ovgu

This module can gather information provided by the webpages of the Otto-von-Guericke University Magdeburg (ovgu). This module currently provides access to the canteen menus and is able to serialize the information to a json string.

Modules

canteen
The ovgu::canteen module is able to parse the ovgu canteen website and provide the information as serializable and deserializable structs.

Enums

Error
The Error enum represents several different error types that are used by results in this crate.