Module ovgu_canteen::ovgu::canteen

The ovgu::canteen module is able to parse the ovgu canteen website and provide the information as serializable and deserializable structs.

Structs

Canteen
A canteen holds all the meals on all available days.

Day
A Day holds all the meals that are available at the given day.

Meal
A Meal holds the meals name, the price, several symbols, additives, and allergenics.

Price
This struct represents the price of a meal.

Enums

Additive
An Additive is used to represent additives of a meal.

Allergenic
This enum represents allergenics that are contained in a meal.

CanteenDescription
This enum identifies a canteen.

Symbol
This enum represents symbols a meal is annotated with.

Traits

FromElement
This trait is used to create an instance from a HTML element reference.