Struct ovgu_canteen::ovgu::canteen::Canteen

pub struct Canteen {
    pub description: CanteenDescription,
    pub days: Vec<Day>,
}

A canteen holds all the meals on all available days.

Fields
description: CanteenDescription

This identifies a canteen.

days: Vec<Day>

All available days holding the meals for this canteen.

Methods

impl Canteen

fn new(desc: CanteenDescription) -> Result<Self, Error>

This method creates a new canteen instance from a given description.

Arguments

desc - The identifier identifying the canteen.

Examples

Create a new canteen:

use ovgu_canteen::ovgu::canteen::{Canteen, CanteenDescription};

let canteen = Canteen::new(CanteenDescription::Downstairs).unwrap();
println!("{}", canteen.days[0].meals[0].name);